JWTAuthorization_put_delete+Till_commentapp

+++ src/blogapp/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render
+
 # from django.http import JsonResponse
 # import json
 from .models import *
@@ -8,7 +9,11 @@ from rest_framework import status
 from django.views.decorators.csrf import csrf_exempt
 from rest_framework.decorators import api_view
 from rest_framework.views import APIView
-from rest_framework.generics import ListCreateAPIView,RetrieveUpdateDestroyAPIView,ListAPIView
+from rest_framework.generics import (
+    ListCreateAPIView,
+    RetrieveUpdateDestroyAPIView,
+    ListAPIView,
+)
 from rest_framework.filters import OrderingFilter
 from rest_framework.exceptions import PermissionDenied
 
@@ -35,7 +40,7 @@ from rest_framework.exceptions import PermissionDenied
 #         # return JsonResponse({"blogs": list(blogs)}, safe=False)
 #         serializer=BlogSerializer(blogs,many=True)
 #         return Response({"blogs":serializer.data})
-    
+
 #     def post(self,request):
 #         serializer=BlogSerializer(data=request.data)
 #         if(serializer.is_valid()):
@@ -66,13 +71,13 @@ from rest_framework.exceptions import PermissionDenied
 #         blog=self.getBlog(id)
 #         if(blog is None):
 #             return Response({"error":"blog is not found."},status=status.HTTP_404_NOT_FOUND)
-        
+
 #         serializer=BlogSerializer(blog,data=request.data,partial=True)
 #         if serializer.is_valid():
 #             serializer.save()
 #             return Response({"message":"Blog updated successfully","blog":serializer.data})
 #         return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
-    
+
 #     def delete(self,request,id):
 #         blog=self.getBlog(id)
 #         if(blog is None):
@@ -82,38 +87,54 @@ from rest_framework.exceptions import PermissionDenied
 #         # return JsonResponse({"message": "Blog deleted successfully"}, status=204)
 #         return Response({"message":"blog deleted successfully"},status=status.HTTP_204_NO_CONTENT)
 
+
 class CreateAndGetAllBlogs(ListCreateAPIView):
-    queryset=Blog.objects.all()
-    serializer_class=BlogSerializer
-    filter_backends=[OrderingFilter] #Enable sorting
-    ordering_fileds=['title','id','create_at','update_at'] #Fields to allow sorting blog
-    oredering=['id'] # Default oredering
+    queryset = Blog.objects.select_related("creator")
+    # queryset=Blog.objects.all()
+    serializer_class = BlogSerializer
+    filter_backends = [OrderingFilter]  # Enable sorting
+    ordering_fileds = [
+        "title",
+        "id",
+        "create_at",
+        "update_at",
+    ]  # Fields to allow sorting blog
+    oredering = ["id"]  # Default oredering
+
 
 class GetUpdateDeleteBlog(RetrieveUpdateDestroyAPIView):
-    queryset=Blog.objects.all()
-    serializer_class=BlogSerializer
+    queryset = Blog.objects.all()
+    serializer_class = BlogSerializer
 
-def perform_update(self,serializer):
-    #Retrieve the existing instance before updating
-    blog=self.get_object()
-    old_image=blog.image #Get the old image instance
-    
-    #Save the new data
-    updated_blog=serializer.save() #Save and retrieve the updated instance
-    
-    #Delete the old image if it has been replaced
+
+def perform_update(self, serializer):
+    # Retrieve the existing instance before updating
+    blog = self.get_object()
+    if blog.creator != self.request.user:
+        raise PermissionDenied("you do not have permission to update this blog.")
+    old_image = blog.image  # Get the old image instance
+
+    # Save the new data
+    updated_blog = serializer.save()  # Save and retrieve the updated instance
+
+    # Delete the old image if it has been replaced
     if old_image and old_image != updated_blog.image:
-        print(f"Deleting old image:{old_image.name}") #Debug log
-        old_image.delete(save=False) #Delete the old image from storage
+        print(f"Deleting old image:{old_image.name}")  # Debug log
+        old_image.delete(save=False)  # Delete the old image from storage
     else:
-        print("No image change detected or no old image to delete.") #Debug log
-        
-def perform_destroy(self,instance):
+        print("No image change detected or no old image to delete.")  # Debug log
+
+
+def perform_destroy(self, instance):
+    if instance.creator != self.request.user:
+        raise PermissionDenied("you do not have permission to delete this blog.")
     if instance.image:
         instance.image.delete(save=False)
     instance.delete()
 
+
 class GetUserBlog(ListAPIView):
-    serializer_class=BlogSerializer
+    serializer_class = BlogSerializer
+
     def get_queryset(self):
         return Blog.objects.filter(creator=self.request.user)
